Split task binding score helper
