+++ apps/desktop/src/main/event-task-state.ts
@@ -4,6 +4,10 @@ import {
   normalizeAgentName,
   normalizeTaskTitle,
 } from './task-run-utils.ts'
+import {
+  findBestIndexedMatch,
+  type BindingHints,
+} from './task-binding-score.ts'
 
 export type TaskStatus = 'queued' | 'running' | 'complete' | 'error'
 
@@ -146,67 +150,6 @@ export function getImmediateParentSession(sessionId?: string | null): string | n
   const parent = sessionLineage.get(sessionId)
   if (!parent || parent === sessionId) return null
   return parent
-}
-
-type BindingHints = {
-  title?: string | null
-  agent?: string | null
-}
-
-function normalizeBindingHints(hints?: BindingHints | null) {
-  return {
-    title: normalizeTaskTitle(hints?.title) || null,
-    agent: normalizeAgentName(hints?.agent) || null,
-  }
-}
-
-function computeBindingScore(
-  candidate: { title?: string | null; agent?: string | null },
-  hints?: BindingHints | null,
-) {
-  const normalizedHints = normalizeBindingHints(hints)
-  if (!normalizedHints.title && !normalizedHints.agent) return 0
-
-  let score = 0
-  const candidateAgent = normalizeAgentName(candidate.agent) || null
-  const candidateTitle = normalizeTaskTitle(candidate.title) || null
-
-  if (normalizedHints.agent && candidateAgent && normalizedHints.agent === candidateAgent) {
-    score += 4
-  }
-
-  if (normalizedHints.title && candidateTitle) {
-    if (normalizedHints.title === candidateTitle) {
-      score += 3
-    }
-  }
-
-  return score
-}
-
-function findBestIndexedMatch<T extends { title?: string | null; agent?: string | null }>(
-  entries: T[],
-  hints?: BindingHints | null,
-) {
-  if (entries.length <= 1) return entries.length === 1 ? 0 : -1
-
-  let bestIndex = -1
-  let bestScore = 0
-  let ambiguous = false
-
-  for (const [index, entry] of entries.entries()) {
-    const score = computeBindingScore(entry, hints)
-    if (score > bestScore) {
-      bestScore = score
-      bestIndex = index
-      ambiguous = false
-    } else if (score > 0 && score === bestScore) {
-      ambiguous = true
-    }
-  }
-
-  if (bestIndex >= 0 && !ambiguous) return bestIndex
-  return -1
 }
 
 function takePendingTaskRunId(parentSessionId: string, hints?: BindingHints | null) {
